+++ test/0_TEST_CASES/2d_compressible_KOmega_flatplate_fixedT_ConstantViscosity.jl
@@ -152,7 +152,7 @@ hardware = Hardware(backend=backend, workgroup=workgroup)
     @test initialise!(model.turbulence.omega, ω_inlet) === nothing
     @test initialise!(model.turbulence.nut, k_inlet/ω_inlet) === nothing
 
-    residuals = run!(model, config)
+    residuals = run!(model, config; progress=false)
 
     inlet = boundary_average(:inlet, model.momentum.U, BCs.U, config)
     outlet = boundary_average(:outlet, model.momentum.U, BCs.U, config)