+++ test/0_TEST_CASES/2d_incompressible_laminar_rotatingFlatplate_MRF.jl
@@ -94,7 +94,7 @@ GC.gc()
 initialise!(model.momentum.U, velocity)
 initialise!(model.momentum.p, 0.0)
 
-residuals = run!(model, config)
+residuals = run!(model, config; progress=false)
 
 MRFCell = 2000 
 inertialCell = 10