+++ test/0_TEST_CASES/2d_incompressible_pitzdaily_Smagorinsky.jl
@@ -125,7 +125,7 @@ initialise!(model.momentum.p, 0.0)
 initialise!(model.turbulence.nut, 0.0)
 
 # Step 11. Run simulation
-residuals = run!(model, config, inner_loops=2, ncorrectors=0, output=OpenFOAM());
+residuals = run!(model, config, inner_loops=2, ncorrectors=0, output=OpenFOAM(), progress=false);
 
 inlet = boundary_average(:inlet, model.momentum.U, BCs.U, config)
 outlet = boundary_average(:outlet, model.momentum.U, BCs.U, config)