+++ test/unit_test_reynolds_stress.jl
@@ -85,7 +85,7 @@ config = Configuration(solvers=solvers, schemes=schemes, runtime=runtime, hardwa
 
 @test initialise!(model.momentum.U, velocity) === nothing
 @test initialise!(model.momentum.p, 0.0) === nothing
-residuals = run!(model, config);
+residuals = run!(model, config; progress=false);
 
 u_rms_exact = U0*A/sqrt(2) 
 v_rms_exact = U0*A/sqrt(2)